[compat,->DEV_IQ_2016] LDRA rule "Signed/unsigned conversion without cast." has been made a countermeasured. refs #3909
git-svn-id: https://tmp.openrtm.org/svn/OpenRTM-aist/branches/DEV_IQ_2016@2941 8bdb766d-82ee-4cb2-b2de-c55db819e4d5

+++ src/lib/coil/common/crc.cpp
@@ -182,7 +182,7 @@ namespace coil
         0xb40bbe37, 0xc30c8ea1, 0x5a05df1b, 0x2d02ef8d,
       };
     unsigned long c(0xffffffffL);
-    for (int i(0); i < len; ++i)
+    for (size_t i(0); i < len; ++i)
       {
         c = crc32tab[(c ^ str[i]) & 0xff] ^ (c >> 8);
       }
